requirements.txt
faiss-cpu
sentence-transformers
pymupdf
requests
beautifulsoup4
youtube-transcript-api
streamlit
langchain_community
tf-keras
